chore(scap): move agent_info to scap_platform
