+++ userspace/libscap/scap-int.h
@@ -50,7 +50,6 @@ struct scap
 
 	uint64_t m_evtcnt;
 	scap_machine_info m_machine_info;
-	scap_agent_info m_agent_info;
 	struct ppm_proclist_info* m_driver_procinfo;
 
 	// Function which may be called to log a debug event
@@ -100,11 +99,6 @@ int32_t scap_check_suppressed(struct scap_suppress *suppress, scap_evt *pevent,
 			      bool *suppressed, char *error);
 
 int32_t scap_procfs_get_threadlist(struct scap_engine_handle engine, struct ppm_proclist_info **procinfo_p, char *lasterr);
-
-//
-// Retrieve agent info.
-//
-void scap_retrieve_agent_info(scap_agent_info* agent_info);
 
 //
 // Retrieve machine info.
