+++ userspace/libscap/scap.h
@@ -70,6 +70,7 @@ typedef struct ppm_evt_hdr scap_evt;
 
 #include "scap_limits.h"
 #include "scap_open.h"
+#include "scap_machine_info.h"
 #include "scap_procs.h"
 
 /* Include engine-specific params. */
@@ -330,16 +331,6 @@ typedef struct _scap_machine_info
 	uint64_t reserved3; ///< reserved for future use
 	uint64_t reserved4; ///< reserved for future use, note: because of scap file captures needs to remain uint64_t, use flags if possible
 }scap_machine_info;
-
-/*!
-  \brief Agent information, not intended for scap file use
-*/
-typedef struct _scap_agent_info
-{
-	uint64_t start_ts_epoch; ///< Agent start timestamp, stat /proc/self/cmdline approach, unit: epoch in nanoseconds
-	double start_time; ///< /proc/self/stat start_time divided by HZ, unit: seconds
-	char uname_r[128]; ///< Kernel release `uname -r`
-}scap_agent_info;
 
 /*!
   \brief Interface address type
@@ -801,15 +792,6 @@ const char* scap_get_ppm_sc_name(ppm_sc_code sc);
   offline captures, it comes from the capture file.
 */
 const scap_machine_info* scap_get_machine_info(scap_t* handle);
-
-/*!
-  \brief Get generic agent information
-
-  \return The pointer to a \ref scap_agent_info structure containing the information.
-
-  \note for live captures only.
-*/
-const scap_agent_info* scap_get_agent_info(scap_t* handle);
 
 /*!
   \brief Set the capture snaplen, i.e. the maximum size an event parameter can
