+++ userspace/libscap/scap_machine_agent.c
@@ -74,58 +74,3 @@ void scap_retrieve_machine_info(scap_machine_info* machine_info, uint64_t boot_t
 	scap_get_bpf_stats_enabled(machine_info);
 #endif
 }
-
-void scap_retrieve_agent_info(scap_agent_info* agent_info)
-{
-	agent_info->start_ts_epoch = 0;
-	agent_info->start_time = 0;
-#ifdef __linux__
-
-	/* Info 1:
-	 *
-	 * Get epoch timestamp based on procfs stat, only used for (constant) agent start time reporting.
-	 */
-	struct stat st = {0};
-	char path[256];
-	snprintf(path, sizeof(path), "/proc/%d/cmdline", getpid());
-	if(stat(path, &st) == 0)
-	{
-		agent_info->start_ts_epoch = st.st_ctim.tv_sec * (uint64_t) SECOND_TO_NS + st.st_ctim.tv_nsec;
-	}
-
-	/* Info 2:
-	 *
-	 * Get /proc/self/stat start_time (22nd item) to calculate subsequent snapshots of the elapsed time
-	 * of the agent for CPU usage calculations, e.g. sysinfo uptime - /proc/self/stat start_time.
-	 */
-	char proc_stat[256];
-	FILE* f;
-	snprintf(proc_stat, sizeof(proc_stat), "/proc/%d/stat", getpid());
-	if((f = fopen(proc_stat, "r")))
-	{
-		unsigned long long stat_start_time = 0; // unit: USER_HZ / jiffies / clock ticks
-		long hz = 100;
-#ifdef _SC_CLK_TCK
-		if ((hz = sysconf(_SC_CLK_TCK)) < 0)
-		{
-			ASSERT(false);
-			hz = 100;
-		}
-#endif
-		if(fscanf(f, "%*d %*s %*c %*d %*d %*d %*d %*d %*lu %*lu %*lu %*lu %*lu %*llu %*llu %*llu %*llu %*d %*d %*d %*lu %llu", &stat_start_time))
-		{
-			agent_info->start_time = (double)stat_start_time / hz; // unit: seconds as type (double)
-		}
-		fclose(f);
-	}
-
-	/* Info 3:
-	 *
-	 * Kernel release `uname -r` of the machine the agent is running on.
-	 */
-
-	struct utsname uts;
-	uname(&uts);
-	snprintf(agent_info->uname_r, sizeof(agent_info->uname_r), "%s", uts.release);
-#endif
-}
